+++ java-io-http/src/main/java/nbbrd/io/http/HttpClient.java
@@ -1,9 +1,11 @@
 package nbbrd.io.http;
 
 import lombok.NonNull;
+import nbbrd.design.NotThreadSafe;
 
 import java.io.IOException;
 
+@NotThreadSafe
 public interface HttpClient {
 
     @NonNull String getDescription();