+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Add ByteCountingClient to HTTP module
+
 ## [0.0.38] - 2026-06-02
 
 ### Changed